+++ smart_media/validators.py
@@ -1,4 +1,5 @@
 from django.conf import settings
+from django.core.validators import FileExtensionValidator
 from django.forms import ValidationError
 from django.template.defaultfilters import filesizeformat
 from django.utils.translation import gettext_lazy as _
@@ -28,3 +29,19 @@ def validate_file_size(data):
         ))
 
     return True
+
+
+class SmartMediaFileExtensionValidator(FileExtensionValidator):
+    """
+    Customized ``FileExtensionValidator`` to directly get the allowed extension from
+    setting ``SMARTIMAGE_ALLOWED_IMAGE_EXTENSIONS``.
+
+    Since allowed extensions are embedded in code from settings, Django won't detect
+    any change model change when changing the setting and so won't create a migration.
+
+    Argument ``allowed_extensions`` is still accepted but is ignored.
+    """
+    def __init__(self, *args, **kwargs):
+        kwargs["allowed_extensions"] = settings.SMARTIMAGE_ALLOWED_IMAGE_EXTENSIONS
+
+        super().__init__(*args, **kwargs)