+++ docs/history.rst
@@ -4,6 +4,16 @@
 History
 =======
 
+Version 0.2.0 - 2023/01/11
+--------------------------
+
+* Added a model field;
+* Added Model admin;
+* Added purge signals;
+* Finished documentation;
+* Finished test coverage;
+
+
 Version 0.1.0 - Unreleased
 --------------------------